promote SimpleMarker and remove Markers
